+++ example/lib/shared/util/enemy_sprite_sheet.dart
@@ -38,7 +38,7 @@ class EnemySpriteSheet {
       );
 
   static SimpleDirectionAnimation get simpleDirectionAnimation =>
-      SimpleDirectionAnimation(
+      SimpleDirectionAnimation.fromFuture(
         idleLeft: idleLeft,
         idleRight: idleRight,
         runLeft: runLeft,